add float commit

+++ task1/src/main/java/average.java
@@ -5,9 +5,9 @@ public class average {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         System.out.println("input 3 numbers");
-        float a = sc.nextInt();
-        float b = sc.nextInt();
-        float c = sc.nextInt();
+        float a = sc.nextFloat();
+        float b = sc.nextFloat();
+        float c = sc.nextFloat();
         float avr = (a + b + c) / 3;
         System.out.println("average=" + avr);
         float half_avr = avr / 2;
